Update docker_build action to pass build_args properly (#15)
Description

This PR aims to fix the build issues with docker image.

RCA:

The build args were not passed properly to the `docker-build-push` action which was resulting in the docker image build failures.
also the `image_name` output was not set properly.

Fix for this commit: 08697bf828221bd35e22d3b24451ab3890df7e96

Signed-off-by: Hasibur Rahman Mohammed <hasiburr@qti.qualcomm.com>